fix:对交叉编译的临时工具链的编译脚本中新增了configure之后的判断逻辑

+++ src/cross_compile_script/libstdc++.sh
@@ -1,4 +1,5 @@
-mkdir -v build && cd build 
+mkdir -v build && pushd build 
+
 ../libstdc++-v3/configure           \
     --host=$LFS_TGT                 \
     --build=$(../config.guess)      \
@@ -7,5 +8,15 @@ mkdir -v build && cd build
     --disable-nls                   \
     --disable-libstdcxx-pch         \
     --with-gxx-include-dir=/tools/$LFS_TGT/include/c++/12.2.0
+
+
+if [ "$?" -eq 1 ];
+then
+    exit $?
+fi
+
 make && make DESTDIR=$LFS install
+
+popd
+
 rm -v $LFS/usr/lib/lib{stdc++,stdc++fs,supc++}.la
